+++ Classes/TiledImage.m
@@ -43,7 +43,7 @@
 
 - (UIImage *)imageAtXIndex:(NSUInteger)xIndex YIndex:(NSUInteger)yIndex {
   if(xIndex >= self.widthCount || yIndex >= self.heightCount) {
-    NSLog(@"Requested out of bounds image: (%i,%i) doesn't exist -- max is (%i, %i)", xIndex, yIndex, self.widthCount, self.heightCount);
+    NSLog(@"Requested out of bounds image: (%lu,%lu) doesn't exist -- max is (%lu, %lu)", (unsigned long)xIndex, (unsigned long)yIndex, (unsigned long)self.widthCount, (unsigned long)self.heightCount);
     return nil;
   }
   NSUInteger tileIndex = TILE_INDEX(xIndex, yIndex);
